Reorganized/clean up components dir
