+++ resources/js/Components/ApplicationInfo.js
