+++ resources/js/Components/ApplicationStatus.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+// import { Link } from 'react-router-dom';
 
 
 class ApplicationStatus extends Component {
