--- resources/js/components/Applications.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-import ApplicationsList from './ApplicationsList';
-import { Link } from 'react-router-dom';
-class Applications extends Component {
-    render() {
-        return (
-            <ApplicationsList />
-        );
-    }
-}
-export default Applications;
