+++ resources/js/components/NewApplication.js
@@ -1,9 +1,7 @@
 import axios from 'axios';
 import React, { Component } from 'react';
-//import Select from 'react-select';
-//import Autosuggest from 'react-autosuggest';
-import InputCompany from './InputCompany';
-import InputLocation from './InputLocation';
+import InputCompany from '@/Components/InputCompany';
+import InputLocation from '@/Components/InputLocation';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
